@@ -39,7 +39,7 @@
     "babel-register": "^6.7.2",
     "chai": "^3.5.0",
     "chokidar-cli": "^1.2.0",
-    "eslint": "5.1.x",
+    "eslint": "6.0.1",
     "mocha": "^2.5.3",
     "proxyquire": "^1.7.4",
     "sinon": "^1.17.3",
